TFP-7026 putter metodene som brukes i opprettForespørsel i ForvaltningBehandlingRestTjeneste for å tydeliggjøre at de brukes i en forvaltningstjeneste
